# openapi.yaml
openapi: "3.1.0"
info:
  title: WSI Viewer Backend
  version: "1.0.0"
  description: |
    Whole-slide image viewer — tile streaming, session management,
    Tumour Content Analysis (TCA), per-slide annotations, and pathology review worklist.
servers:
  - url: http://localhost:8080
    description: Local dev
...
components:
...
  schemas:
...
    Point:
      type: object
      required: [x, y]
      properties:
        x: { type: number, minimum: 0, maximum: 1 }
        y: { type: number, minimum: 0, maximum: 1 }

    # ── Annotations ───────────────────────────────────────────────────────
    PolygonAnnotation:
      type: object
      properties:
        id:         { type: string }
        vertices:   { type: array, items: { $ref: '#/components/schemas/Point' } }
        label:      { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    LabelAnnotation:
      type: object
      properties:
        id:         { type: string }
        anchor:     { $ref: '#/components/schemas/Point' }
        offset:     { $ref: '#/components/schemas/Point' }
        text:       { type: string }
        created_at: { type: string, format: date-time }
    MeasureAnnotation:
      type: object
      properties:
        id:           { type: string }
        start:        { $ref: '#/components/schemas/Point' }
        end:          { $ref: '#/components/schemas/Point' }
        distance_um:  { type: number, nullable: true }
        distance_px:  { type: number, nullable: true }
        created_at:   { type: string, format: date-time }
    SlideAnnotations:
      type: object
      properties:
        slide_key: { type: string }
        polygons:  { type: array, items: { $ref: '#/components/schemas/PolygonAnnotation' } }
        labels:    { type: array, items: { $ref: '#/components/schemas/LabelAnnotation' } }
        measures:  { type: array, items: { $ref: '#/components/schemas/MeasureAnnotation' } }

    # ── TCA ───────────────────────────────────────────────────────────────
    TCAAggregates:
      type: object
      properties:
        overall_tumour_pct:    { type: number, nullable: true }
        overall_stroma_pct:    { type: number, nullable: true }
        overall_necrosis_pct:  { type: number, nullable: true }
        hotspot_tumour_pct:    { type: number, nullable: true }
        hotspot_row:           { type: integer, nullable: true }
        hotspot_col:           { type: integer, nullable: true }
        total_cells_analysed:  { type: integer }
        tissue_cells:          { type: integer }
    TCADensityBand:
      type: object
      properties:
        label:      { type: string }
        min_pct:    { type: number }
        max_pct:    { type: number }
        cell_count: { type: integer }
        area_um2:   { type: number, nullable: true }
    TCAGridCell:
      type: object
      properties:
        row:              { type: integer }
        col:              { type: integer }
        tumour_pct:       { type: number }
        stroma_pct:       { type: number }
        necrosis_pct:     { type: number, nullable: true }
        inflammation_pct: { type: number, nullable: true }
        other_pct:        { type: number, nullable: true }
    TCAAnalysisResult:
      type: object
      properties:
        slide_key:     { type: string }
        available:     { type: boolean }
        analysed_at:   { type: string, format: date-time, nullable: true }
        aggregates:    { $ref: '#/components/schemas/TCAAggregates', nullable: true }
        density_bands: { type: array, items: { $ref: '#/components/schemas/TCADensityBand' } }
        grid_cells:    { type: array, items: { $ref: '#/components/schemas/TCAGridCell' } }
    TCASummary:
      type: object
      properties:
        slide_key:            { type: string }
        available:            { type: boolean }
        analysed_at:          { type: string, format: date-time, nullable: true }
        overall_tumour_pct:   { type: number, nullable: true }
        overall_stroma_pct:   { type: number, nullable: true }
        overall_necrosis_pct: { type: number, nullable: true }
        hotspot_tumour_pct:   { type: number, nullable: true }
        total_cells_analysed: { type: integer }
    TCAGridResponse:
      type: object
      properties:
        slide_key:  { type: string }
        available:  { type: boolean }
        grid_cells: { type: array, items: { $ref: '#/components/schemas/TCAGridCell' } }
        total:      { type: integer }

    # ── Worklist ──────────────────────────────────────────────────────────
...
paths:
...
  /api/tca/{slide_key}:
    get:
      tags: [tca]
      summary: Full TCA analysis result
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
          description: Filename stem (e.g. 1087-25) or full filename
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TCAAnalysisResult' } } } }
        "500": { description: Internal analysis error }
  /api/tca/{slide_key}/summary:
    get:
      tags: [tca]
      summary: Lightweight TCA aggregate summary
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TCASummary' } } } }
  /api/tca/{slide_key}/grid:
    get:
      tags: [tca]
      summary: Spatial grid cells for heatmap rendering
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TCAGridResponse' } } } }
  /api/tca/{slide_key}/cache:
    delete:
      tags: [tca]
      summary: Invalidate cached TCA analysis
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Cache cleared }

  # ── Annotations ──────────────────────────────────────────────────────────
  /api/annotations/{slide_key}:
    get:
      tags: [annotations]
      summary: Get all annotations for a slide
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SlideAnnotations' } } } }
    delete:
      tags: [annotations]
      summary: Clear all annotations for a slide
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Cleared }
  /api/annotations/{slide_key}/polygons:
    post:
      tags: [annotations]
      summary: Add polygon annotation
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vertices]
              properties:
                vertices: { type: array, items: { $ref: '#/components/schemas/Point' }, minItems: 3 }
                label:    { type: string, nullable: true }
                id:       { type: string, nullable: true }
      responses:
        "201": { description: Created }
  /api/annotations/{slide_key}/polygons/{annotation_id}:
    patch:
      tags: [annotations]
      summary: Update polygon
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
        - name: annotation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Updated }
    delete:
      tags: [annotations]
      summary: Delete polygon
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
        - name: annotation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /api/annotations/{slide_key}/labels:
    post:
      tags: [annotations]
      summary: Add label annotation
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "201": { description: Created }
  /api/annotations/{slide_key}/labels/{annotation_id}:
    patch:
      tags: [annotations]
      summary: Update label
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
        - name: annotation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Updated }
    delete:
      tags: [annotations]
      summary: Delete label
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
        - name: annotation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /api/annotations/{slide_key}/measures:
    post:
      tags: [annotations]
      summary: Add measure annotation
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
      responses:
        "201": { description: Created }
  /api/annotations/{slide_key}/measures/{annotation_id}:
    delete:
      tags: [annotations]
      summary: Delete measure
      parameters:
        - name: slide_key
          in: path
          required: true
          schema: { type: string }
        - name: annotation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }

  # ── Worklist ──────────────────────────────────────────────────────────────
...